Integrating server. Change background
