+++ client/src/Components/BannerTopBar/index.js
@@ -7,7 +7,7 @@ const BannerTopBar = ({
   title,
   text,
 }) => (
-  <div className={`bannerTopBar ${size}`} style={backgroundUrl ? { backgroundImage: `url(http://aacorporatesitedevelop.azurewebsites.net/img/photos/modules.jpg)` } : {}}>
+  <div className={`bannerTopBar ${size}`} style={backgroundUrl ? { backgroundImage: `url(https://www.advancedalgos.net/img/photos/financial-beings.jpg)` } : {}}>
     <div className='caption'>
       <Typography className='title' variant='h1' align='center'>{title}</Typography>
       <Typography className='text' variant='h2' align='center'>{text}</Typography>
