+++ client/src/Components/Main/Tabs/Browse.js
@@ -1,4 +1,6 @@
 import React from 'react';
+import { Query } from 'react-apollo';
+import { FINANCIAL_BEINGS } from "../../../GraphQL/Calls";
 import { Typography } from '@material-ui/core';
 
 import { withStyles } from '@material-ui/core/styles';
@@ -16,6 +18,21 @@ class History extends React.Component {
         gutterBottom
       >
       Browse financial beings
+
+        <Query query={FINANCIAL_BEINGS}>
+          {({ loading, error, data }) => {
+            if (loading) return 'Loading...';
+            if (error) return `Error! ${error.message}`;
+            const list = data.financialBeings.map((event, index) => (
+              <p>{event.name}</p>
+            ));
+            return (
+              <React.Fragment>
+                { list }
+              </React.Fragment>
+            );
+          }}
+        </Query>
       </Typography>
     );
   }
