+++ client/src/Components/Main/Tabs/index.js
@@ -1,4 +1,4 @@
-import Search from './Search';
+import Search from './Search/Search';
 import Genealogy from './Genealogy';
 import Browse from './Browse';
