+++ client/src/GraphQL/Calls/index.js
@@ -1,0 +1,36 @@
+import gql from 'graphql-tag';
+
+
+export const FINANCIAL_BEINGS = gql`
+  query FB {
+  financialBeings {
+    id
+    type
+    kind
+    name
+    slug
+    avatar
+    creator
+    status {
+      status
+      reason
+    }
+    updatedOn
+    createdOn
+    team
+    admins {
+      adminId
+    }
+  }
+}
+`;
+
+export const FB_BY_NAME = gql`
+query financialBeingsByPartialName {
+  financialBeingsByPartialName(name: "2") {
+    id
+    name
+    kind
+  }
+}
+`;
